Automatically try to set upstream branch, fixes weird sync edge cases
When the user manually sets the sync remote, git does not know the
upstream branch, and any existing upstream branch info is cleared.

This change tries to determine and set the correct upstream branch on
sync to make syncing succeed with a new remote.

If an empty remote is set, nothing is pulled and the local tree is
pushed and created on the remote. If the remote is not empty, it sets
the upstream branch and attempts to pull in its tree.

Of course this could be manually set and configured by the user, but
with prs we try to do as much automatically as possible to prevent
annoying prompts and difficult instructions for new users.