Fix some command-line flag constraints
The generate command now always required the copy flag, because the
timeout option requires it, which was always set because of a default
value in clap. This removes the default value from clap, and moves it
into the argument matcher.

The default value for the store argument has also been moved to the
matcher to prevent future issues.