Disable git SSH connection reuse for now
This disables the git SSH connection reuse feature, because it causes
failures for some git hosts. This can be worked around, but requires
some additional logic. This commit disables the feature until this is
properly implemented, just so we're on the safe side.

See: https://github.com/timvisee/prs/issues/5#issuecomment-803940880