Fix error when initializing Tomb when having a large store
The logic for determining the size of a password store did not work when
the password store is not a Tomb. This was used to determine the initial
Tomb size. The default size is low, causing an error when the user's
password store is large. This has now been fixed.